refactor code. add working window version
