+++ TronGame/settings.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#define SCR_WIDTH 600
+#define SCR_HEIGHT 480
+
+#define DELAY 0.01
+
+#define BG_TEXT "data/images/background.jpg"
+#define FONT "data/gameFont.ttf"
